Clean IvyThing.scala; move ProxyFromEnvTest.
