+++ amm/runtime/src/main/scala/ammonite/runtime/tools/IvyThing.scala
@@ -4,8 +4,6 @@ import java.io.PrintWriter
 
 import ammonite.util.{Printer, Util}
 
-import scala.language.postfixOps
-
 
 object IvyConstructor extends IvyConstructor
 trait IvyConstructor{
@@ -35,7 +33,6 @@ object IvyThing{
       logger.init()
       Some(logger)
     }
-
 
     val start = coursier.Resolution(dependencies.toSet)
 
@@ -87,3 +84,4 @@ object IvyThing{
               .mkString(".")
   
 }
+
