+++ amm/src/test/scala/ammonite/main/ProxyFromEnvTest.scala
@@ -1,4 +1,4 @@
-package ammonite.runtime.tools
+package ammonite.main
 
 import utest._
 /**
